+++ Bank/CurrencyConverter.cs
@@ -10,9 +10,9 @@ namespace Bank
     {
         public static readonly Currency MainCurrency = Currency.PLN;
 
-        public static Dictionary<Currency, double> Prices = new Dictionary<Currency, double>();
+        public static Dictionary<Currency, int> Prices = new Dictionary<Currency, int>();
 
-        public static double Convert(Currency from, Currency to, double amount)
+        public static double Convert(Currency from, Currency to, int amount)
         {
             if (Prices.ContainsKey(from)
               && Prices.ContainsKey(to))
